requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
SQLAlchemy>=2.0.25
psycopg2-binary>=2.9.9
pydantic>=2.7.0
pydantic-settings>=2.2.1
python-dotenv>=1.0.1
python-jose[cryptography]>=3.3.0
